GroupRepository

+++ src/WebApi/Repositories/GroupRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using MongoDB.Driver;
 using WebApi.Common;
 using WebApi.Models.Mongodb;
@@ -19,30 +20,34 @@ namespace WebApi.Repositories
 
         public List<Group> GetAllGroupsByUserId(string userId)
         {
+            NullCheck(userId, nameof(userId));
+
             var filter = Builders<Group>.Filter.Eq(UserIdField, userId);
             return _groups.Find(filter).ToList();
         }
 
         public GroupResult InsertGroup(string groupName, string userId)
         {
-            var group = new Group
-            {
-                GroupName = groupName,
-                UserId = userId
-            };
+            NullCheck(groupName, nameof(groupName));
+            NullCheck(userId, nameof(userId));
 
             var groupResult = new GroupResult
             {
-                Group = group
+                Group = new Group
+                {
+                    GroupName = groupName,
+                    UserId = userId
+                }
             };
 
             if (IsExisted(groupName, userId))
             {
                 groupResult.Result = Result.Exist;
+                groupResult.Reason = $"{groupName} is existed";
             }
             else
             {
-                _groups.InsertOne(group);
+                _groups.InsertOne(groupResult.Group);
                 groupResult.Result = Result.Succeed;
             }
 
@@ -51,20 +56,23 @@ namespace WebApi.Repositories
 
         public GroupResult UpdateGroup(string newGroupName, string groupId, string userId)
         {
-            var group = new Group
-            {
-                GroupId = groupId,
-                UserId = userId
-            };
+            NullCheck(newGroupName, nameof(newGroupName));
+            NullCheck(groupId, nameof(groupId));
+            NullCheck(userId, nameof(userId));
 
             var groupResult = new GroupResult
             {
-                Group = group
+                Group = new Group
+                {
+                    GroupId = groupId,
+                    UserId = userId
+                }
             };
 
             if (IsExisted(newGroupName, groupId, userId))
             {
                 groupResult.Result = Result.Exist;
+                groupResult.Reason = $"{newGroupName} is existed";
             }
             else
             {
@@ -72,39 +80,65 @@ namespace WebApi.Repositories
                 var update = Builders<Group>.Update.Set(GroupNameField, newGroupName);
                 _groups.UpdateOne(filter, update);
                 groupResult.Result = Result.Succeed;
-                group.GroupName = newGroupName;
+                groupResult.Group.GroupName = newGroupName;
             }
             return groupResult;
         }
 
         public GroupResult GetGroup(string groupId, string userId)
         {
+            NullCheck(groupId, nameof(groupId));
+            NullCheck(userId, nameof(userId));
+
+            var groupResult = new GroupResult();
             var filter = Builders<Group>.Filter.Eq(GroupIdField, groupId) & Builders<Group>.Filter.Eq(UserIdField, userId);
             var group = _groups.Find(filter).FirstOrDefault();
-            return new GroupResult
+            if (group != null)
             {
-                Group = group,
-                Result = Result.Succeed
-            };
+                groupResult.Group = group;
+                groupResult.Result = Result.Succeed;
+            }
+            else
+            {
+                groupResult.Result = Result.Failed;
+                groupResult.Reason = "Can not find such group";
+            }
+            return groupResult;
         }
 
         public GroupResult DeleteGroup(string groupId, string userId)
         {
-            var filter = Builders<Group>.Filter.Eq(GroupIdField, groupId) & Builders<Group>.Filter.Eq(UserIdField, userId);
-            _groups.DeleteOne(filter);
-            return new GroupResult
+            NullCheck(groupId, nameof(groupId));
+            NullCheck(userId, nameof(userId));
+
+            var queryResult = GetGroup(groupId, userId);
+            var groupResult = new GroupResult();
+            if (queryResult.Result == Result.Exist)
             {
-                Group = new Group
+                var filter = Builders<Group>.Filter.Eq(GroupIdField, groupId) & Builders<Group>.Filter.Eq(UserIdField, userId);
+                _groups.DeleteOne(filter);
+                groupResult.Group = queryResult.Group;
+            }
+            else
+            {
+                groupResult.Group = new Group
                 {
                     GroupId = groupId,
                     UserId = userId
-                },
-                Result = Result.Succeed
-            };
+                };
+                groupResult.Result = Result.Failed;
+                groupResult.Reason = "Can not find such group";
+            }
+
+
+            return groupResult;
         }
 
         public bool IsExisted(string groupName, string userId)
         {
+            NullCheck(groupName, nameof(groupName));
+            NullCheck(userId, nameof(userId));
+
             var filter = Builders<Group>.Filter.Eq(UserIdField, userId) & Builders<Group>.Filter.Eq(GroupNameField, groupName);
             var group = _groups.Find(filter).FirstOrDefault();
             return group != null;
@@ -112,10 +146,22 @@ namespace WebApi.Repositories
 
         public bool IsExisted(string groupName, string groupId, string userId)
         {
+            NullCheck(groupName, nameof(groupName));
+            NullCheck(groupId, nameof(groupId));
+            NullCheck(userId, nameof(userId));
+
             var filter = Builders<Group>.Filter.Eq(GroupIdField, groupId) & Builders<Group>.Filter.Eq(UserIdField, userId) &
                          Builders<Group>.Filter.Eq(GroupNameField, groupName);
             var group = _groups.Find(filter).FirstOrDefault();
             return group != null;
+        }
+
+        private static void NullCheck(string filed, string filedName)
+        {
+            if (string.IsNullOrEmpty(filed))
+            {
+                throw new NullReferenceException($"{filedName} is null");
+            }
         }
     }
 }
